Keep Windows Dongl helpers local
