+++ simpleble/CMakeLists.txt
@@ -466,8 +466,7 @@ if(SIMPLEBLE_TEST)
         ${CMAKE_CURRENT_SOURCE_DIR}/test/src/main.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/test/src/test_utils.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/test/src/test_bytearray.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/test/src/test_buffer_overflow.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/test/src/test_usb_helper_windows.cpp)
+        ${CMAKE_CURRENT_SOURCE_DIR}/test/src/test_buffer_overflow.cpp)
     set_target_properties(simpleble_test PROPERTIES
         CXX_VISIBILITY_PRESET hidden
         VISIBILITY_INLINES_HIDDEN YES
